refactor: move model pkg to internal
