+++ handler/register_handler_test.go
@@ -1,7 +1,7 @@
 package handler
 
 import (
-	"basic_server/model"
+	"basic_server/internal/model"
 	"basic_server/repository/mocks"
 	"bytes"
 	"net/http"
