+++ repository/mocks/user_repository_mock.go
@@ -1,7 +1,7 @@
 package mocks
 
 import (
-	"basic_server/model"
+	"basic_server/internal/model"
 	"basic_server/repository"
 	"errors"
 )
