+++ server/repository/post_repository.go
@@ -1,7 +1,7 @@
 package repository
 
 import (
-	"basic_server/model"
+	"basic_server/internal/model"
 
 	"gorm.io/gorm"
 )
