+++ server/service/post_service.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"basic_server/model"
+	"basic_server/internal/model"
 
 	"gorm.io/gorm"
 )
